manifests: move groups from deployer
The `deployer` subpackages (api, rte, sched) grow a nice abstraction
over the `manifest` package, creating and handing a group of related
manifests the clients can use to setup a component.

This code actually belongs to `manifests` (not to `deployer`)
so in this patch we move the code, fixing all the references
along the way.
