+++ manifests
@@ -1,1 +1,1 @@
-pkg/manifests/manifests
+pkg/manifests/yaml
