+++ pkg/manifests/yaml/rte/clusterrole.yaml
