+++ pkg/manifests/yaml/rte/daemonset.yaml
