+++ pkg/manifests/yaml/rte/namespace.yaml
