+++ pkg/manifests/yaml/rte/serviceaccount.yaml
