+++ pkg/manifests/yaml/sched/clusterrole-allinone.yaml
