+++ pkg/manifests/yaml/sched/clusterrole.yaml
