+++ pkg/manifests/yaml/sched/clusterrolebinding-vol-sched.yaml
