+++ pkg/manifests/yaml/sched/deployment.yaml
