+++ pkg/manifests/yaml/sched/namespace.yaml
