+++ pkg/manifests/yaml/update.sh
